// src/api/gen/types/TxDtoV1.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AddressAliasAdd } from './AddressAliasAdd.ts'
import type { AddressAliasRemove } from './AddressAliasRemove.ts'
import type { AuthorityAdd } from './AuthorityAdd.ts'
import type { AuthorityRemove } from './AuthorityRemove.ts'
import type { NetworkParamsSet } from './NetworkParamsSet.ts'
import type { TokenBurn } from './TokenBurn.ts'
import type { TokenCreate } from './TokenCreate.ts'
import type { TokenMint } from './TokenMint.ts'
import type { TokenUpdate } from './TokenUpdate.ts'
import type { ValidatorAdd } from './ValidatorAdd.ts'
import type { ValidatorRemove } from './ValidatorRemove.ts'
import type { Vote } from './Vote.ts'

export const txDtoV1VersionEnum = {
  V1: 'V1',
} as const

export type TxDtoV1VersionEnumKey =
  (typeof txDtoV1VersionEnum)[keyof typeof txDtoV1VersionEnum]

export const txDtoV1TypeEnum = {
  TRANSFER: 'TRANSFER',
  BIP_CREATE: 'BIP_CREATE',
  BIP_VOTE: 'BIP_VOTE',
} as const

export type TxDtoV1TypeEnumKey =
  (typeof txDtoV1TypeEnum)[keyof typeof txDtoV1TypeEnum]

export const txDtoV1NetworkEnum = {
  MAINNET: 'MAINNET',
  TESTNET: 'TESTNET',
} as const

export type TxDtoV1NetworkEnumKey =
  (typeof txDtoV1NetworkEnum)[keyof typeof txDtoV1NetworkEnum]

export const txDtoV1PayloadTypeEnum = {
  BIP_ADDRESS_ALIAS_ADD: 'BIP_ADDRESS_ALIAS_ADD',
  BIP_ADDRESS_ALIAS_REMOVE: 'BIP_ADDRESS_ALIAS_REMOVE',
  BIP_AUTHORITY_ADD: 'BIP_AUTHORITY_ADD',
  BIP_AUTHORITY_REMOVE: 'BIP_AUTHORITY_REMOVE',
  BIP_NETWORK_PARAMS_SET: 'BIP_NETWORK_PARAMS_SET',
  BIP_TOKEN_BURN: 'BIP_TOKEN_BURN',
  BIP_TOKEN_CREATE: 'BIP_TOKEN_CREATE',
  BIP_TOKEN_MINT: 'BIP_TOKEN_MINT',
  BIP_TOKEN_UPDATE: 'BIP_TOKEN_UPDATE',
  BIP_VOTE: 'BIP_VOTE',
  BIP_VALIDATOR_ADD: 'BIP_VALIDATOR_ADD',
  BIP_VALIDATOR_REMOVE: 'BIP_VALIDATOR_REMOVE',
} as const

export type TxDtoV1PayloadTypeEnumKey =
  (typeof txDtoV1PayloadTypeEnum)[keyof typeof txDtoV1PayloadTypeEnum]

export type TxDtoV1 = {
  /**
   * @type string | undefined
   */
  hash?: string
  /**
   * @type string | undefined
   */
  version?: TxDtoV1VersionEnumKey
  /**
   * @type string | undefined, date-time
   */
  timestamp?: string
  /**
   * @type string | undefined
   */
  type?: TxDtoV1TypeEnumKey
  /**
   * @type string | undefined
   */
  network?: TxDtoV1NetworkEnumKey
  /**
   * @type integer | undefined, int64
   */
  nonce?: number
  /**
   * @type string | undefined
   */
  sender?: string
  /**
   * @type string | undefined
   */
  recipient?: string
  /**
   * @type string | undefined
   */
  amount?: string
  /**
   * @type string | undefined
   */
  fee?: string
  /**
   * @type string | undefined
   */
  tokenAddress?: string
  /**
   * @type string | undefined
   */
  message?: string
  /**
   * @type string | undefined
   */
  referenceHash?: string
  /**
   * @type string | undefined
   */
  signature?: string
  /**
   * @type string | undefined
   */
  payloadType?: TxDtoV1PayloadTypeEnumKey
  payload?:
    | AddressAliasAdd
    | AddressAliasRemove
    | AuthorityAdd
    | AuthorityRemove
    | NetworkParamsSet
    | TokenBurn
    | TokenCreate
    | TokenMint
    | TokenUpdate
    | ValidatorAdd
    | ValidatorRemove
    | Vote
  /**
   * @type integer | undefined, int32
   */
  size?: number
  /**
   * @type string | undefined
   */
  blockHash?: string
  /**
   * @type integer | undefined, int64
   */
  blockHeight?: number
  /**
   * @type integer | undefined, int32
   */
  index?: number
}